Arreglos en cronometrador y ensamblador y fabrica

+++ src/aplicacion/Cronometrador.java
@@ -7,14 +7,14 @@ public class Cronometrador extends Thread {
 
     private Semaphore semContMutEx;//Semaforo para establecer exclusion mutua entre el cronometrador y el gerente
     private int cont;//contador
-    private int duracionDia;//tiempo en segundos que dura un dia en el programa
+    private double duracionDia;//tiempo en segundos que dura un dia en el programa
     private int diasDespacho;//cantidad de dias entre despachos
     private JTextField statusCTextField;//TextField donde se muestra el status del cronometrador
     private JTextField diasDespachoTextField;//TextField donde se muestra los dias que faltan para el despacho
     private long tiempoTrabajo;//tiempo que tarda el cronometrador en disminuir el contador en milisegundos para poder hacer el primer thread.sleep
     private long tiempoDescanso;//tiempo que duerme el cronometrador al dia
 
-    public Cronometrador(Semaphore semContMutEx, int cont, int duracionDia, int diasDespacho, JTextField statusCTextField, JTextField diasDespachoTextField) {
+    public Cronometrador(Semaphore semContMutEx, int cont, double duracionDia, int diasDespacho, JTextField statusCTextField, JTextField diasDespachoTextField) {
         this.semContMutEx = semContMutEx;
         this.cont = cont;
         this.duracionDia = duracionDia;
